backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pdfplumber
pytesseract
Pillow
pandas
numpy
scipy
ta
yfinance
newsapi-python
transformers
torch
hmmlearn
scikit-learn
pgmpy
statsmodels
arch
shap
lime
plotly
matplotlib
ruptures
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-core
langgraph
chromadb
sentence-transformers
pytorch-forecasting
sdv
python-multipart
httpx
groq
